Completely rewritten the data loading model, making use of the ViewAdapter to load each custom GraphView view by view rather than all at once directly from the Room database using ViewModel (ViewModel is still used, but only to get a reference the entire database). Also ironed out a lot of creases to ensure that data loads correctly. (i.e loads correctly no matter which way you scroll through the RecyclerView // generates a new view when data is entered beyond the scope of the current view).
Also improved the readabilty of snackbar for all theme (Wisteria was previously unreadable) by setting textcolor from the main activity.

NOTE: In this commit the "Yearly" view is actually set to minutely to give you a faster impression of how the app handles user input.

This will most likely be one of my final commits for this app. I'm aware the code could be made more readable / concise in many ways, particularly though integration of the code which displays the current debt value (Currently making use of shared preferences) with the DebtValue objects stored in the database. This is a result of the iterative development of the app, and also my lack of knowledge on good data storage practices in the early stages of the project. In future projects I will attempt to minimise these sorts of issues by planning ahead and applying the up to date techniques which I am now much more familiar with. I'm eager to move on to the next project and am satisfied with the current end results.

Furthermore, I have yet to test this app out on any devices other than my own phone and so there may be unexpected results on other devices emulators. I've followed the compatibility guidelines as far as I'm aware, but there may well be some issues that I'm not currently aware of. This is something I will test out before I move on to my next project.